Big refactor of project structure
