--- backend/app/admin.py
@@ -1,11 +1,0 @@
-"""
-This file controls the administrative interface for the sonification web app.
-"""
-
-from django.contrib import admin
-
-models_to_register = [
-]
-
-for model in models_to_register:
-    admin.site.register(model)
