--- backend/app/models.py
@@ -1,3 +1,0 @@
-"""
-Models for the the sonification web app.
-"""
