+++ backend/app/tests/test_data/black.jpg
