+++ backend/app/tests/test_data/blue.jpg
