+++ backend/app/tests/test_data/test_image.jpg
