+++ backend/app/tests/test_data/white.jpg
